implement: fragment support graph refactor + bugfix (#212)
